Module dim::core

Module contains our core initialization logic.

Statics

METADATA_PATH

Path to where metadata is stored and should be fetched to.

Functions

run_scanners

Function dumps a list of all libraries in the database and starts a scanner for each which monitors for new files using fsnotify. It also scans all orphans on boot.

warp_core

Type Definitions

DbConnection

EventTx

StateManager