Module dim::external::filename

Module contains a common interface for extracting and obtaining filename metadata.

Structs

CombinedExtractor

A special filename metadata extractor that combines torrent_name_parser and anitomy which in some cases is necessary. TNP is really good at extracting show titles but not season and episode numbers. Anitomy excels at this. Here we combine the title extracted by TPN and the season and episode number extracted by Anitomy.

Metadata

Traits

FilenameMetadata