Module dim::routes

Contains all of the routes exposed by the webapi. This module contains all the routes supported by the API.

Errors

All routes can error out. If some unexpected behavior has occured, the server will return a non 2XX HTTP code, alongside the error and error description encoded with the following format:

{
  "error": "<error type>",
  "message": "<error message>"
}

Common errors

DatabaseError - All routes can spurriously fail when making a database transaction, or trying to acquire a database connection. When this happens DatabaseError will be returned.

Modules

auth

This module contains all docs and APIs related to authentication and user creation.

dashboard

general

host

This module contains the docs and implementation of various host-related API endpoints.

invites

This module contains various routes used to deal with invite tokens.

library

media

mediafile

rematch_media

settings

statik

stream

tv

user

This module contains all docs and APIs related to users and user metadata.