Module dim::routes::auth

This module contains all docs and APIs related to authentication and user creation.

Request Authentication and Authorization

Most API endpoints require a valid JWT authentication token. If no such token is supplied, the API will return Unauthenticated. Authentication tokens can be obtained by logging in with the login method. Authentication tokens must be passed to the server through a Authroization header.

Example of an authenticated call

curl -X POST http://127.0.0.1:8000/api/v1/auth/whoami -H "Content-type: application/json" -H
"Authorization: eyJhb....."

Token expiration

By default tokens expire after exactly two weeks, once the tokens expire the client must renew them. At the moment renewing the token is only possible by logging in again.

Modules

filters

Functions

admin_exists

login

POST /api/v1/auth/login

register

POST /api/v1/auth/register