Function dim::routes::auth::register

pub async fn register(
    new_user: Login, 
    conn: DbConnection
) -> Result<impl Reply, DimError>

POST /api/v1/auth/register

Method will create a new user and return it a authentication token if a user has been successfuly created.

Request

This method accepts a JSON body that deserializes into Login. If there are no other users in the database, this route will give the new user owner permissions. Additionally this route will not require an invite token.

If there is a user in the database, this request will require an invite token and the user will be given only user permissions.

Example

curl -X POST http://127.0.0.1:8000/api/v1/auth/login -H "Content-type: application/json" -d
'{"username": "testuser", "password": "testpassword", "invite_token":
"72390330-b8af-4413-8305-5f8cae1c8f88"}'

Response

If a user is successfully created, this method will return status 200 0K as well as the create user’s username.

{
  "username": "...."
}

Errors

NoToken - Either the request doesnt contain an invite token, or the invite token is invalid.