Function dim::routes::invites::get_all_invites

pub async fn get_all_invites(
    conn: DbConnection,
    user: User
) -> Result<impl Reply, DimError>

GET /api/v1/auth/invites

Method will retrieve and return all invite tokens in the database.

Authorization

This route requires a valid authentication token to be supplied. The token must have owner permissions.

Request

Example

curl -X GET http://127.0.0.1:8000/api/v1/auth/invites -H "Authorization: ...."

Response

The route will return a response with the following schema

[
  {
    "id": String,
    "created": i64,
    "claimed_by": Option<String>,
  },
  ...
]

Example

[
  {
    "id": "079a38b4-d39f-4a9e-9a18-964f225b75d3",
    "created": 1638708402,
    "claimed_by": "admin"
  },
  {
    "id": "844caa7b-f54f-a9ea-4444-555555555555",
    "created": 1640000000,
  }
]

Errors

Unauthorized - Returned if the authentication token lacks owner permissions