Function dim::routes::library::library_get

pub async fn library_get(
    conn: DbConnection,
    _user: User
) -> Result<impl Reply, DimError>

GET /api/v1/library

Method maps to GET /api/v1/library and returns a list of all libraries in te database.

Request

An authenticated request must be made to this endpoint. No other inputs are required.

Example

curl -X GET http://127.0.0.1:8000/api/v1/library -H "Authroization: ...."

Response

This method will return 200 OK as well as a JSON payload of the following format:

[
  {
    "id": number,
    "name": string,
    "media_type": "movie" | "tv",
    "media_count": number,
  },
  ...
]