Module dim::routes::media

Modules

filters

Functions

delete_media_by_id

Method mapped to DELETE /api/v1/media/<id> is used to delete a media entry for the library. ONly authenticated users can query this.

get_media_by_id

Method mapped to GET /api/v1/media/<id> returns info about a media based on the id queried. This method can only be accessed by authenticated users.

get_media_files

get_mediafile_tree

GET /api/v1/media/<id>/tree

map_progress

Method mapped to POST /api/v1/media/<id>/progress is used to map progress for a certain media to the user. This is useful for remembering progress for a movie etc.

tmdb_search

Method mapped to GET /api/v1/media/tmdb_search is used to quickly search TMDB based on 3 params, one of which is optional. This is used client side in the rematch utility

update_media_by_id

Method mapped to PATCH /api/v1/media/<id> is used to edit information about a media entry manually. It is used in the web ui to manually edit metadata of a media.