Function dim::routes::media::map_progress

pub async fn map_progress(
    conn: DbConnection, 
    id: i64, 
    offset: i64, 
    user: User
) -> Result<impl Reply, DimError>

Method mapped to POST /api/v1/media/<id>/progress is used to map progress for a certain media to the user. This is useful for remembering progress for a movie etc.

Arguments

id - id of the media to modify

Query params

offset - offset in seconds