Function dim::routes::media::update_media_by_id

pub async fn update_media_by_id(
    id: i64, 
    data: UpdateMedia, 
    _user: User, 
    conn: DbConnection
) -> Result<impl Reply, DimError>

Method mapped to PATCH /api/v1/media/<id> is used to edit information about a media entry manually. It is used in the web ui to manually edit metadata of a media.

Arguments

conn - database connection

id - id of the media we want to edit

data - the info that we changed about the media entry

_user - Auth middleware