Function dim::routes::mediafile::rematch_mediafile

pub async fn rematch_mediafile(
    conn: DbConnection, 
    mediafiles: Vec<i64>, 
    external_id: String, 
    media_type: String
) -> Result<impl Reply, DimError>

Method mapped to PATCH /api/v1/mediafile/match used to match a unmatched(orphan) mediafile to a tmdb id.

Arguments

conn - database connection

log - logger

event_tx - websocket channel over which we dispatch a event notifying other clients of the new metadata

mediafiles - ids of the orphan mediafiles we want to rematch

tmdb_id - the tmdb id of the proper metadata we want to fetch for the media