Function dim::routes::user::change_password

pub async fn change_password(
    conn: DbConnection,
    user: User,
    old_password: String,
    new_password: String
) -> Result<impl Reply, DimError>

POST /api/v1/user/password

Method changes the password for a logged in account.

Request

This method accepts a JSON body with the following schema:

{
  "old_password": String,
  "new_password": String,
}

The old_password field in the JSON payload must be the currently registered password for this user. The new_password field is the new password that we want to set.

Example

curl -X POST http://127.0.0.1:8000/api/v1/user/password -H "Content-type: application/json"
-H "Authroization: ..." -d '{"old_password": "testPass", "new_password": "newTestPass"}'

Response

If the password is successfully changed, the method will simply return 200 0K.

Errors

InvalidCredentials - The provided old_password is incorrect or the authentication token is invalid.