Function dim::streaming::ffcheck

pub fn ffcheck() -> Vec<Result<Box<str>, &'static str>>

ffcheck - Check if “ffmpeg” and “ffprobe” are accessable through std::process::Command.

This will run ffmpeg -version and ffprobe -version and return a vec of the stdout output if successfull or the binaries name if not.

Example

use streaming::ffcheck;

for result in ffcheck() {
    match result {
        Ok(stdout) => println!("{:?}", stdout),
        Err(program) => eprintln!("Failed to get the `-version` output of {:?}", program),
    }
}