Function dim::routes::user::whoami

pub async fn whoami(
    user: User, 
    conn: DbConnection
) -> Result<impl Reply, DimError>

GET /api/v1/user

Method returns metadata about the currently logged in user.

Request

This method takes in no additional parameters or data.

Authorization

This method requires a valid authentication token.

Example

curl -X GET http://127.0.0.1:8000/api/v1/user -H "Authorization: ..."

Response

This method will return a JSON payload with the following schema:

{
  "picture": Option<String>,
  "spentWatching": i64,
  "username": String,
  "roles": [String]
}

Example

{
  "picture": "/images/avatar.jpg",
  "spentWatching": 12,
  "username": "admin",
  "roles": ["owner"],
}