Trait dim::scanner::MediaMatcher

pub trait MediaMatcher: Send + Sync {
    fn batch_match<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        tx: &'life1 mut Transaction<'life2>,
        provider: Arc<dyn ExternalQueryIntoShow>,
        work: Vec<WorkUnit>
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait;
    fn match_to_id<'life0, 'life1, 'life2, 'life3, 'async_trait>(
        &'life0 self,
        tx: &'life1 mut Transaction<'life2>,
        provider: Arc<dyn ExternalQueryIntoShow>,
        work: WorkUnit,
        external_id: &'life3 str
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        'life3: 'async_trait;
}

Trait that must be implemented by a media matcher. Matchers are responsible for fetching their own external metadata but it is provided a metadata provider at initialization time.

Required Methods

fn batch_match<'life0, 'life1, 'life2, 'async_trait>(
&'life0 self,
tx: &'life1 mut Transaction<'life2>,
provider: Arc<dyn ExternalQueryIntoShow>,
work: Vec<WorkUnit>
) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,

fn match_to_id<'life0, 'life1, 'life2, 'life3, 'async_trait>(
&'life0 self,
tx: &'life1 mut Transaction<'life2>,
provider: Arc<dyn ExternalQueryIntoShow>,
work: WorkUnit,
external_id: &'life3 str
) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
'life3: 'async_trait,

Match a WorkUnit to a specific external id.